Add contract module
